+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# rquery 1.3.5 2019/06/25
+
+ * Remove str2lang() from vignette (method is not in older R)
+ * Improve expression re-mapping (include n()).
+
 # rquery 1.3.4 2019/06/14
 
  * Work on schema qualifications.